Use GLib.timeout_add instead of GLib.timeout_add_seconds [VPNLINUX-478]
The reason is that the latter groups timers together so that they fire at
the same time. In the context of the reconnector, we don't want that.
We want that events from sources like network manager run as soon as
possible to avoid timing issues.

For example, one of these timing issues was occurring when the
reconnection was happening just after resuming from suspend.
Even though the reconnection was being scheduled to run in ~ 1
second after the VPN drop was detected, sometimes it run
before the network manager event in charge of flagging the
connection as inactive (after the VPN drop).

https://lazka.github.io/pgi-docs/#GLib-2.0/functions.html#GLib.timeout_add_seconds