Do not use NetworkManager to detect network connectivity [VPNLINUX-568]
After enabling the IPv6 leak protection implementation based on the dummy network
interface, the network monitor stopped working, which impacted VPN reconnections.

Up until now, the network monitor was based on NetworkManager. Network manager
was notifying our network monitor whenever there was a network connectivity
state change. However, after introducing the dummy interface to protect against
IPv6 leaks, NetworkManager didn't notify the network monitor anymore of network
connectivity state changes.

The reason was that, once the VPN connection was down, the interface with more
priority was always the dummy interface, which always reports network
connectivity. Therefore, there were no connection state changes. And, because
of this, the reconnector was blindly retrying to connect to the VPN, since
the network monitor was reporting that there was network connectivity.

After long discussions, we agreed to actively poll for network connection
state changes, without relying on NetworkManager.

At this point, coming up with another sort of kill switch is not a viable
solution as it would delay the public release even further.