Remove IPv6 leak protection when quitting the app [VPNLINUX-568]
When the VPN state machine was in error state, we were not
removing the IPv6 leak protection when quitting the app.

To fix this, we now always disconnect when the VPN widget
is "unrealized", forcing the VPN state machine to transition
from error state to disconnected state, which is the state
in charge of removing IPv6 leak protection.