Fix server list refresh [VPNLINUX-524]
Up until now, the app was checking every 60 minutes if the server list
was expired and if it was, it recreated the whole server list.

Now we follow the standard intervals to refresh the server list and
the server loads independently. When the server loads expired,
then only the loads are updated but the whole list is not recreated.