Switch to AsyncExecutor [VPNLINUX-688]
Up until now, every time we run async code from the Gtk main loop we started a new
asyncio event loop via a ThreadPoolExecutor. This caused thread-safety issues
because async code is not supposed to be run simultaneously by multiple threads.

This commit introduces the AsyncExecutor, which replaces the ThreadPoolExecutor.
The AsyncExecutor has a single thread running the asyncio loop and always delegates
execution of coroutine functions to the same asyncio loop. This gets rid of
thread safety issues because the execution of asyncio code is always delegated
to a single thread, in a thread-safe manner.

In addition, the AsyncExecutor also allows running regular functions that are not
asyncio coroutines. In this case, the execution is delegated to a regular
ThreadPoolExecutor.