Add in-tree Flatpak support
I took the Flatpak manifest from
https://github.com/flathub/com.protonvpn.www and converted it to
JSON for this, so that it can be used by GNOME Builder for local
Flatpak builds. Credits to @proletarius101 for the original work!

GNOME Builder only supports the JSON format.
Builder is the recommended IDE for developers building GTK and
GNOME apps. It has built-in support for Flatpak builds with just a
click. https://wiki.gnome.org/Newcomers/BuildProject

Making it easy for the GNOME/GTK community to build your app can
attract valuable contributors.

In the future, if there's interest, this same Flatpak manifest could
be used by a GitHub Action to produce a Flatpak bundle out of a
CI job.