This simplifies versioning the repo. [VPNLINUX-806]
Instead of having to manually update the debian and rpm changelogs everytime
we do a release we update a 'versions.yml' file which contains everything we
need to generate the changelogs for both systems.

This change also changes how we version release candidates. Instead of
using the 0.x system in rpm, we now use the ~x suffix throughout.

setup.py now also uses versions.yml to know the latest version of the
project.

We introduce a new file versions.yml which has the version information.
We introduce some new utility scripts for helping, the most important
being:
- validate_versions.py : Which checks the versions.yml is well formed.
- build_packages.py    : Which should be run when versions.yml is
  changed.
- versions.py	       : Is the python module which contains all the
  logic.