[VPNLINUX-884] Only allow kill switch settings modifications when disconnected from the VPN
The kill switch blocks VPN traffic if it's enabled when a connection is already active.
Given that the routing table kill switch implementation behaves differently on openvpn
and wireguard, it's easier to only allow the user to change KS settings after disconnecting
from the VPN. That's what's being added by this commit.